docs(setqflist): clarify that there is a fourth `action` for `setqflist`: ' '
When reading the docs for `setqflist`, one would see the three
possible values for `action` in a list format and decide which
option should be passed. Confusingly there is a fourth option,
passing " " or nothing at all, which creates a new quickfix list
on the quickfix list stack. This option could even be considered
the default, but it appears as an afterthought in the docs.

Move the paragraph explaining the " " option to the list with
the other options.