fix(lsp): set tabSize from 'shiftwidth', not 'softtabstop'
The use of 'softtabstop' to set tabSize was introduced in 5d5b068,
replacing 'tabstop'.  This was a step in the right direction, but if we
look past the name tabSize and at the actual purpose of the field,
it's the indentation width used when formatting.  (The field should have
been called "indentSize", but there's nothing to be done about that
now.)  This corresponds to the Vim option 'shiftwidth', not
'softtabstop'.  The latter has the comparatively mundane purpose of
controlling what happens when you hit the tab key (and even this is
incomplete, as it fails to account for 'smarttab').