fix: correct event order and values for reusing the NoName buffer
Editing a real file after opening nvim without a file to edit reuses the
"NoName" buffer that nvim uses as a placeholder. Previously the buffer
creation code would emit BufDelete and BufWipeout events manually for
the NoName buffer and then call buf_freeall a bit later with that
buffer. buf_freeall emits BufUnload and optionally emits BufDelete and
BufWipeout via flags that indicate to emit them. The buf_freeall call
for the NoName buffer wasn't given those flags, so it only emitted
BufUnload.

All of this means that when reusing the NoName buffer, we'd emit
BufDelete, BufWipeout, and then BufUnload. The docs state that BufUnload
is before BufDelete, so that's one bug. A second bug was that the buffer
filename was set to the newly edited file's name in the meantime, so the
BufUnload event had the new file in <afile> when it's called, not the empty
string.

This pulls the call to buf_freeall up to where the delete and wipeout
events were manually emitted. It uses buf_freeall to emit those events,
so removes that duplication. That gets the events to go in the correct
unload, delete, and wipeout order and corrects the filename for the
unload event.