test(script_spec): use source over writing to file
This reverts the change made in #16064
(https://github.com/neovim/neovim/commit/72652cbc46f568128bfc296ba63fb2d26941da8e),
as it's not needed.

The reason why this fails when helpers.exec is used directly, but not when using
helpers.source, is due to source calling write_file, which applies a second
dedent to remove the indent before the heredoc END marker (required to make
heredocs work). Though helpers.source is now deprecated, and uses exec under the
hood, it still calls dedent to mimic the old behaviour, so we can still use it.