vim-patch:8.2.2675: directory change in a terminal window shell is not followed
Problem:    Directory change in a terminal window shell is not followed.
Solution:   Add the 'autoshelldir' option. (closes vim/vim#6290)
https://github.com/vim/vim/commit/8b9abfd86c736ed3f298dfa8b50a962c118b3983

Note that Vim's libvterm has a slightly different API - in particular,
the vterm fallbacks take a VTermStringFragment instead of a char*+len,
which means the parsing of OSC sequences had to be tweaked a little bit.