feat(terminal): support OSC7 ('autoshelldir')
This commit introduces a new 'autoshelldir' setting controlling whether
Neovim should aknowledge OSC7 sequences emitted by the programs running
in its terminal buffer.

The OSC7 sequence is sent by the child process to indicate that its
current working directory changed.

When Neovim receives OSC7, it stores the received path in a buffer-local
variable named "terminal_job_cwd", which is available to the user. It
also stores the path as the terminal buffer's name. This enables
'autochdir' to work as expected with buffer terminals.

Note that Vim also has a setting named 'autoshelldir' that does
something similar. This setting was implemented in patch 8.2.2675
(Vim commit 8b9abfd86c736ed3f298dfa8b50a962c118b3983). However, Vim's
approach has several drawbacks causing possible desynchronization
between the child process' cwd and the parent window. This is why this
patch isn't a port of Vim's feature, but a new implementation. However,
in the hope of sharing code with Vim in the future, an attempt was made
to keep function names and interfaces similar to Vim's. Moreover, the
url_decode function was imported as-is.