feat(lsp): support dynamic registration of didChangeConfiguration
Add client support for server to dynamically register and unregister the
workspace/didChangeConfiguration handler.

The main motiviation for implementing this feature is that currently
neovim (or rather nvim-lspconfig [0]) sends the didChangeConfiguration
notification just after initialization to pass client settings to the
server regardless whether the server has registered this capability or
not. This is the first step to fix this.

[0]: https://github.com/neovim/nvim-lspconfig/blob/21102d5e3b6ffc6929d60418581ac1a29ee9eddd/lua/lspconfig/configs.lua#L175