fix: don't mutate opts parameter of vim.keymap.del
`vim.keymap.del` takes an `opts` parameter that lets caller refer to and
delete buffer-local mappings. For some reason the implementation of
`vim.keymap.del` mutates the table that is passed in, setting
`opts.buffer` to `nil`. I'm not sure why it does this since `opts` is
never again accessed/used by the implementation anyway.

This seems wrong and is also undocumented. This bit me because I was
setting and then later unsetting a collection of mappings for some
buffer, and I created a local `opts = {buffer=my_buffer_number}`, and
then I got "no such mapping" errors.