feat(treesitter): make revolving of injection `@language` configurable
The current language injection engine allows capturing an
injection language from a document. In future, we would be more flexible
about that and also allow resolving abbreviations like `cxx` for `cpp`
or `py` for `python` or allow to configure more ambiguous injections
that could be named in captures like `@pythondocs` where the user could
decide whether that identifier should correspond to an preferred injection
language. Also sometimes the context of the injection like the files in
a certain project could require different decisions on the injection
logic.

An overwritable  function would allow to have possibly complex logic
in user land while core could live with a 1:1 mapping for now and
maybe an extension using the Lua filetype module to resolve a language
by filetype extension.