feat(runtime): specify when deprecated functions will be removed
This is primarily intended to act as documentation for the developer so
they know exactly when and what to remove. This will help prevent the
situation of deprecated code lingering for far too long as developers
don't have to worry if a function is safe to remove.

This could've been achieved just as well by a simple comment before the
function, but thought this information could be useful to end users as
well, which is why I opted to add the information to the deprecation
message itself.