feat(lsp): add vim.lsp.config
Add a configuration frontend to easily configure LSP servers on a
per-filetype basis. This is essentially a thin wrapper around the
following process:

1. Create a FileType autocommand for the given filetype
2. When the autocommand fires, determine the "root" for the given buffer
   using the root markers passed to vim.lsp.config()
3. If a client already exists for this filetype and root directory,
   attach the buffer to that client
4. If no client exists, create a new client configuration table using
   the options passed to vim.lsp.config() and start a new client with
   vim.lsp.start_client()

This also exposes a new vim.lsp.buf_start() function that can be used to
manually start LSP clients for servers configured with vim.lsp.config()
when the "autostart" option is false.