ci(uncrustify): lint codebase with uncrustify
This lint job will ensure that the C codebase is properly formatted at
all times. This will help us get rid of the automatable parts of clint as well
as reducing a lot of noise from the dev-style section in the documentation.

Opted to manually compile uncrustify and cache the resulting binary
instead of downloading it from homebrew to save CI time.