ci(clint): remove character width warning
The warning for exceeding character width is more of an annoyance than
any benefit. The only time it seems to trigger is when a line actually
needs to exceed the limit, such as an URL or a long string message.
Furthermore, both uncrustify and clang-format is already very good at
automatically containing the code at a certain width, rendering manual
checks unnecessary.