fix: always close libuv handles on exit
When EXITFREE is defined Nvim does not automatically close libuv
handles, which causes a ~2 second delay when Nvim exits as it waits for
the handles to be closed. It is very common for libuv handles to still
be open when Nvim exits, as plugins use them and don't always close them
(particularly with timers which are re-used and are thus
never closed). Now that EXITFREE is defined in Debug builds by default,
this causes a 2 second delay upon each exit which quickly becomes
annoying.

Instead, Nvim should *always* close existing libuv handles, regardless
of whether or not EXITFREE is defined.